+++ include/nova_renderer/util/platform.hpp
