+++ include/nova_renderer/util/result.hpp
