+++ src/loading/shaderpack/shaderpack_data.cpp
@@ -6,7 +6,7 @@
 #include "../../util/vma_usage.hpp"
 
 #include "../json_utils.hpp"
-#include "shaderpack_data.hpp"
+#include "../../../include/nova_renderer/shaderpack_data.hpp"
 
 namespace nova::renderer {
     /*!
