+++ src/loading/shaderpack/shaderpack_loading.hpp
@@ -7,8 +7,8 @@
 #define NOVA_RENDERER_SHADERPACK_LOADING_HPP
 
 #include <future>
-#include "../../util/filesystem.hpp"
-#include "shaderpack_data.hpp"
+#include "../../../include/nova_renderer/util/filesystem.hpp"
+#include "../../../include/nova_renderer/shaderpack_data.hpp"
 
 namespace nova::ttl {
     class task_scheduler;
