+++ src/loading/zip_folder_accessor.cpp
@@ -5,7 +5,7 @@
 
 #include "zip_folder_accessor.hpp"
 #include "../util/logger.hpp"
-#include "../util/utils.hpp"
+#include "../../include/nova_renderer/util/utils.hpp"
 
 #include <array>
 #include <memory>
