+++ src/render_engine/dx12/dx12_texture.hpp
@@ -8,7 +8,7 @@
 
 #include <d3d12.h>
 #include <wrl.h>
-#include "../../loading/shaderpack/shaderpack_data.hpp"
+#include "../../../include/nova_renderer/shaderpack_data.hpp"
 
 using Microsoft::WRL::ComPtr;
