+++ src/render_engine/vulkan/auto_allocating_buffer.hpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 #include <vulkan/vulkan.h>
-#include "../../util/result.hpp"
+#include "../../../include/nova_renderer/util/result.hpp"
 #include "cached_buffer.hpp"
 
 namespace nova::renderer {
