+++ src/render_engine/vulkan/swapchain.hpp
@@ -10,7 +10,7 @@
 #include <glm/glm.hpp>
 #include <vector>
 #include <vulkan/vulkan.h>
-#include "../../util/utils.hpp"
+#include "../../../include/nova_renderer/util/utils.hpp"
 
 namespace nova::renderer {
     class vulkan_render_engine;
