+++ src/render_engine/vulkan/vulkan_type_converters.hpp
@@ -6,7 +6,7 @@
 #define NOVA_RENDERER_VULKAN_TYPE_CONVERTERS_HPP
 
 #include <vulkan/vulkan.h>
-#include "../../loading/shaderpack/shaderpack_data.hpp"
+#include "../../../include/nova_renderer/shaderpack_data.hpp"
 
 #ifdef Always // RIP X11 and its macros
 #undef Always
