+++ src/render_engine/vulkan/vulkan_utils.cpp
@@ -4,7 +4,7 @@
  */
 
 #include "vulkan_utils.hpp"
-#include "../render_engine.hpp"
+#include "../../../include/nova_renderer/render_engine.hpp"
 
 namespace nova::renderer {
     std::string vk_result_to_string(VkResult result) {
