+++ src/render_engine/vulkan/x11_window.cpp
@@ -2,7 +2,7 @@
 // Created by jannis on 01.09.18.
 //
 
-#include "../../platform.hpp"
+#include "../../../include/nova_renderer/util/platform.hpp"
 
 #ifdef NOVA_LINUX
