+++ src/util/result.cpp
@@ -1,4 +1,4 @@
-#include "result.hpp"
+#include "../../include/nova_renderer/util/result.hpp"
 #include "fmt/format.h"
 
 namespace nova::renderer {
