+++ src/util/windows.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../platform.hpp"
+#include "../../include/nova_renderer/util/platform.hpp"
 
 #ifndef WIN32_LEAN_AND_MEAN
 	#define WIN32_LEAN_AND_MEAN
